requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
numpy
pandas
scikit-learn
fastapi-cors
sqlalchemy
psycopg2-binary
python-dotenv